Added Models and Camera Script
Added a new C# script to make the camera follow a designated player
ship. (Though maybe this should be done by the game controller?)
Imported the models and put them into the test scene as entities to play
around with. Added a RR_Models and RR_Materials folder for models and
materials. Added a particle effect and sphere child to the Resonator
entity to see how it would look.