Added status indicator circles to our unit prefabs.
Added status indicator circles to the unit prefabs, as well as a script
to manipulate the unit prefabs and textures to use to color/fill in the
materials.

Defects: There was a defect where the material references weren't being
properly set in the prefabs and would instead turn into a bright magenta
placeholder. Fixed.

There were a number of defects with BroadcastMessage where the messages
were not received because the messages broadcast did not match the
actual name of the methods intended to be called. Fixed.