Added Background Music & BGMPlayer.
3:40 PM - 4:40 PM: Added a BGMPlayer to the testing scene. Also added
two music track assets that are free to use under Creative Commons Zero
license and updated the Readme.md to credit that asset use.

Defect Added/Resolved: Originally had a direct reference to an
AudioSource in the script instead of pulling it from connected assets.
Added in design, fixed in code. Accidentally forgot to capitalize the
'get' in 'GetComponent<T>()'.  Originally tried to directly set a value
in the BGM player by using GetComponent outside of any of Unity's
Events. Simply declared that same variable, but assign it when
UpdateAudioSource is called, and UpdateAudioSource is called in the
Start() event and whenever the audio is paused or unpaused.

Pausing the game also causes music to pause and unpausing causes music
to resume.