WARNING: BROKEN - Units now use rigidbodies.
Tweaking things to use rigidbodies and addforce to move around so that
they stop passing through eachother. Though- as you'll see if you run
this version- that comes with a lot of interesting defects that has to
do with the way our existing player controls ignores forces and the way
that units don't factor in the Y axis while rigidbodies does. Also,
there's a warning message about computing mesh inertia tensors.