Fixed resonator spawning bug, plus other stuff
Fixed the resonator spawning bug. It was a line of code that should have
been commented out but wasn't. Also redid most of the variables in
GameController.cs into properties. To access them, just do
yourReferenceToTheScript.PropertyYouWantToAccess. Properties always
begin with capital letters. I've updated various scripts to account for
this (getting rid of the script.GetBlah() methods and mathod calls).
Also worked a bit on the GUI popup windows.