New Resonators Sorta Work! Yay!
The new Resonator in the Res test scene now works! Mostly! The
indicators don't respond properly on generic units, but does fine on the
new Resonator itself. If you want to use the new Resonator in the
default test scene, you must change the Game Controller in the inspector
to point to the new Resonator instead of the old one.